Factor out `limit` logic into its own module
